fixed issue #339
Former-commit-id: cd5ee3381495a72af88e27d21c931d25f8c22739

+++ cocos2d/Layer.m
@@ -151,7 +151,7 @@
 	return [[[self alloc] initWithColor:color] autorelease];
 }
 
-- (id) initWithColor:(ccColor4B)color width:(GLint)w  height:(GLint) h
+- (id) initWithColor:(ccColor4B)color width:(GLfloat)w  height:(GLfloat) h
 {
 	if( (self=[super init]) ) {
 		color_.r = color.r;
